polish some tests
